Working data transfer and LCD display
I have the LCD working, the pyserial program now correctly transmits data.
It turns out the issue was not with the atmega after all.
I needed to flush the input and output buffers of pySerial

Also, sending the temperature: it would sometimes accidentially send the wind speed as the temperature
The problem was the time.sleep(5) method.

For some reason that offset the cycle or messed up the buffer in some way.

Anyway, I used .read(1) on the pyserial side and transmitByte('y') on the atmega side
to make sure the two were in sync.
However, I have since made it so there should be no "write_timeout" for writing with pyserial
so theoretically this technique is no longer needed.

However, for now, I am leaving it in.

I wrote a function to display the temperature and the wind speed on the LCD but it does not work.
I am pushing it with this commit, but will change it and update later if I can.

~Logan Haser